リアルタイム株価チェック処理追加、イベント、リスナ追加、マイグレーションデータ型修正
	modified:   app/Console/Kernel.php
	new file:   app/Events/MinitlyStocksCheck.php
	modified:   app/Http/Controllers/RealtimeController.php
	new file:   app/Listeners/UpdateStocksInfo.php
	modified:   app/Providers/EventServiceProvider.php
	modified:   database/migrations/2020_05_13_203541_create_realtime_settings_table.php
	modified:   database/migrations/2020_05_13_205637_create_realtime_checkings_table.php
	modified:   database/migrations/2020_05_15_133922_create_stocks_table.php
	modified:   resources/views/realtime_checking.blade.php
	modified:   routes/web.php

+++ database/migrations/2020_05_13_203541_create_realtime_settings_table.php
@@ -17,9 +17,9 @@ class CreateRealtimeSettingsTable extends Migration
             $table->id();
             $table->integer('user_id');
             $table->integer('stock_id');
-            $table->integer('upperlimit')->nullable();  //上限値
+            $table->float('upperlimit')->nullable();  //上限値
             $table->dateTime('upperlimit_settingat')->nullable();
-            $table->integer('lowerlimit')->nullable();  //下限値
+            $table->float('lowerlimit')->nullable();  //下限値
             $table->dateTime('lowerlimit_settingat')->nullable();
             $table->float('changerate')->nullable();    //変化率
             $table->dateTime('changerate_changerate')->nullable();
