+++ database/migrations/2020_05_13_205637_create_realtime_checkings_table.php
@@ -16,9 +16,9 @@ class CreateRealtimeCheckingsTable extends Migration
         Schema::create('realtime_checkings', function (Blueprint $table) {
             $table->id();
             $table->integer('realtime_setting_id');
-            $table->integer('price')->nullable();   //現在値
+            $table->float('price')->nullable();   //現在値
             $table->dateTime('price_checkingat')->nullable();   //現在値
-            $table->integer('pre_price')->nullable();   //前回(1分前)現在値
+            $table->float('pre_price')->nullable();   //前回(1分前)現在値
             $table->dateTime('pre_price_checkingat')->nullable();   //前回(1分前)現在値
             $table->float('rate')->nullable();   //前回(1分前)現在値
             $table->dateTime('rate_checkingat')->nullable();   //前回(1分前)現在値
